Paging init, needs to be implemented

+++ app/src/main/java/com/projects/mercadopago/data/network/MercadoPagoApiService.kt
@@ -13,7 +13,7 @@ interface MercadoPagoApiService {
         @Query("q")
         query: String,
         @Query("offset")
-        offset: Double,
+        offset: Int,
     ): ResponseModel
 }
